+++ test/index.ts
@@ -1,1 +1,3 @@
-console.log({ "import.meta.url": import.meta.url } as any);
+// WARN: do not use these words as key
+console.log({ import_meta_url: import.meta.url } as any);
+console.log({ filename: __filename } as any);